wip #4 setting up vector pair for ..
use in not using word already used during an active search. Need to figure out how to clear entire vector each time. could just do a for loop but I'm trying to see if there's a way to use the clear command to do it instead for easier to read code.