wip#5 after work on vector_pair need to make map...
instead so it fits in the single if statement. working on the map and how to clear it after each bad word search is done but it seems to have some of the same issues with vector.clear where the memory of the vector is still there if accessed even if size of vector is down since it's not deallocated from clear (from how I currently understand it) and it's undefined behavior and could still be the previous value or a random garbage value.