wip #2 doesn't work with test case no node with...
node->val == val it does however apparently traverse the whole tree since it returns 7 which is the last tree_node it should traverse. need to find a way to return NULL 
without an error popping up where it says returning null is wrong since it's not a tree_node struct.